Add handle deeplink on IOS
