+++ src/commonMain/kotlin/io/github/jan/supabase/annotiations/SupabaseExperimental.kt
@@ -1,0 +1,4 @@
+package io.github.jan.supabase.annotiations
+
+@RequiresOptIn(level = RequiresOptIn.Level.ERROR, message = "This API is experimental and may not be stable yet")
+annotation class SupabaseExperimental
